requirements.txt
# Python 3.12.2 is recommended

google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
numpy==2.3.1
openai==1.97.0
pillow==11.3.0
pytest==7.4.4
paddleocr==3.1.0
paddlex==3.1.3
pygame==2.6.1
python-dotenv==1.1.1
PyYAML==6.0.2
qwen-agent==0.0.27
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
SpeechRecognition==3.14.3
websocket-client==1.8.0
bilibili-api-python
edge_tts
image-hash==4.3.2
